Add null ScriptLine test and consolidate ScriptLine tests
